move hero bottom
